% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wally-package.R
\docType{data}
\name{divat}
\alias{divat}
\title{divat data}
\format{A subsample consisting of 1300 observations on the following 3 variables.
\describe{ \item{pi}{5-year risk prediction of kidney graft failure.}
\item{status}{0=censored, 1=kidney graft failure}
\item{time}{time to event (i.e., time to kidney graft failure or loss of follow-up)}}}
\description{
Extracted data from a french population based cohort (DIVAT cohort). The dataset includes
followup information on kidney graft failure  outcome and predicted 5-year risks based on
based on the subject specific information which includes age, gender,
cardiovascular and diabetes histories, monitoring of the evolution of the kidney function
measured via serum creatinine and relevant characteristics of his or her kidney donor.
Graft failure is defined as either death with functioning kidney graft or return to dialysis.
The prediction model from which the predictions have been computed has been previously fitted
using an independent training sample from the DIVAT data. Details about data and modeling can
be found in Fournier et al. (2016).
}
\examples{
data(divat)
}
\references{
Fournier, M. C., Foucher, Y., Blanche, P., Buron, F., Giral, M., & Dantan, E. (2016).
A joint model for longitudinal and time-to-event data to better assess the specific
role of donor and recipient factors on long-term kidney transplantation outcomes.
European journal of epidemiology, 31(5), 469-479.
}
\keyword{datasets}

